Rename ffmpeg_capture.py to video_capture.py
