+++ scripts/asset_processor/video_capture.py
